Move requires_api_version to helpers
Because it's a helper, not a base thing. In preparation for
moving/deleting the unit test base classes.
