+++ tests/integration/build_test.py
@@ -8,7 +8,7 @@ import six
 
 from docker import errors
 
-from ..base import requires_api_version
+from ..helpers import requires_api_version
 from .base import BaseIntegrationTest
