+++ tests/integration/service_test.py
@@ -2,7 +2,7 @@ import random
 
 import docker
 
-from ..base import requires_api_version
+from ..helpers import requires_api_version
 from .base import BaseIntegrationTest
