+++ tests/integration/swarm_test.py
@@ -1,7 +1,7 @@
 import docker
 import pytest
 
-from ..base import requires_api_version
+from ..helpers import requires_api_version
 from .base import BaseIntegrationTest
